Upload the project.
This update con my raytracing visual studio 2010 express project, all
deps, and my starting main file. This will change over time. I'm using
SDL1 for the framebuffer so it possible to port this project to mac,
linux, and because I'm creating it in such a way that it won't depend on
the underlying operating system. The only thing you'll need to do is
compile it for you're platform. I'll probably start using cmake
eventually.